+++ rust/cli/src/main.rs
@@ -130,9 +130,9 @@ struct Experimental {
     #[arg(hide = true, long, default_value = "1")]
     batch_size: usize,
 
-    /// Number of tasks for batch parallelism.
-    #[arg(hide = true, long, default_value = "1")]
-    num_tasks: usize,
+    /// Number of tasks for batch parallelism (defaults to the number of CPUs).
+    #[arg(hide = true, long)]
+    num_tasks: Option<usize>,
 
     /// Number of threads for graph parallelism (ONNX Runtime configuration).
     ///
@@ -157,7 +157,8 @@ struct Experimental {
 async fn main() -> Result<()> {
     let flags = Arc::new(Flags::parse());
     ensure!(0 < flags.experimental.batch_size, "--batch-size cannot be zero");
-    ensure!(0 < flags.experimental.num_tasks, "--num-tasks cannot be zero");
+    let num_tasks = flags.experimental.num_tasks.unwrap_or_else(num_cpus::get);
+    ensure!(0 < num_tasks, "--num-tasks cannot be zero");
     ensure!(
         flags.path.iter().filter(|x| x.to_str() == Some("-")).count() <= 1,
         "only one path can be the standard input"
@@ -168,11 +169,9 @@ async fn main() -> Result<()> {
     if flags.colors.disable {
         colored::control::set_override(false);
     }
-    let (result_sender, mut result_receiver) = tokio::sync::mpsc::channel::<Result<Response>>(
-        flags.experimental.num_tasks * flags.experimental.batch_size,
-    );
-    let (batch_sender, batch_receiver) =
-        async_channel::bounded::<Batch>(flags.experimental.num_tasks);
+    let (result_sender, mut result_receiver) =
+        tokio::sync::mpsc::channel::<Result<Response>>(num_tasks * flags.experimental.batch_size);
+    let (batch_sender, batch_receiver) = async_channel::bounded::<Batch>(num_tasks);
     tokio::spawn({
         let flags = flags.clone();
         let result_sender = result_sender.clone();
@@ -183,7 +182,7 @@ async fn main() -> Result<()> {
         }
     });
     let magika = Arc::new(build_session(&flags)?);
-    for _ in 0..flags.experimental.num_tasks {
+    for _ in 0..num_tasks {
         tokio::spawn({
             let magika = magika.clone();
             let batch_receiver = batch_receiver.clone();